+++ compare-frontend/src/app/state/actions.ts
@@ -22,5 +22,6 @@ export const autoPilotDequeue = createAction('Auto-pilot: dequeuing', props<{ sc
 
 export const mapMove = createAction('Map move', props<{ scenario: ScenarioName, partition: Partition, zoom: number, center: number[] }>());
 export const mapClick = createAction('Map click', props<{ scenario: ScenarioName, partition: Partition, location: number[] | undefined, clickedFeature?: {compositeId: string} }>());
+export const mapLayerOpacity = createAction('Map layer opacity', props<{ scenario: ScenarioName, partition: Partition, layerCompositeId: string, opacity: number }>());
 
 export const togglePartition = createAction('Toggling partition', props<{ scenario: ScenarioName, partition: Partition }>());